first rev separating core api
