+++ stashy/__init__.py
@@ -1,6 +1,15 @@
-__version__ = "0.1"
+from .core import Core
+from .client import StashClient
 
-from .client import Stash
+class Stash(object):
+    _url = "/"
+
+    def __init__(self, base_url, username=None, password=None, verify=True, session=None):
+        self._client = StashClient(base_url, username, password, verify, session=session)
+
+        # init sub-clients
+        self.core = Core(base_url, self._client._session)
+
 
 def connect(url, username, password, verify=True):
     """Connect to a Stash instance given a username and password.
@@ -10,5 +19,3 @@ def connect(url, username, password, verify=True):
     verifcation.
     """
     return Stash(url, username, password, verify)
-
-__all__ = ['connect']
